made the corrections
