+++ src/pages/donate-now/DonateNow.jsx
@@ -17,7 +17,7 @@ import usePageView from '../../hooks/usePageView';
 import { DONATION, postData } from '../../services/ApiClient';
 import toast from 'react-hot-toast';
 import useDonationPrompt from '../../hooks/useDonationPrompt';
-import Cookies from 'js-cookie';
+import useExtractURLParams from '../../hooks/useExtractURLParams';
 const UTM_CAMPAIGN_SOURCE = 'utm_source';
 const UTM_REFERRER = 'utm_referrer';
 const PHONE_NUMBER_REGEX = /^(\+[1-9]{1}[0-9]{3,14})?([0-9]{9,14})$/;
@@ -35,16 +35,7 @@ export const DonateNow = () => {
   const { redirectModal, handleRedirectModal } = useContext(ModalContext) || {};
   const { fullname, email, phone, company, toggleAnonymous, invalidPhoneNumber, focus, title, message, error, errorMessage } = userData;
   const { currentText, nextText, swapText, setSwapText } = useDonationPrompt(amount);
-
-  const getReferralParams = () => {
-    const referralParams = Cookies.get('referralParams');
-    if (referralParams) {
-      return JSON.parse(referralParams);
-    }
-    return null;
-  };
-
-  const paramsObject = getReferralParams();
+  const { referralObject } = useExtractURLParams();
 
   /**
    * On success or On failure, flutterwave redirects the user to the specified route with two params attached
@@ -141,10 +132,10 @@ export const DonateNow = () => {
       redirect_url: urlWithoutParams,
       payment_options: 'card',
       currency: 'NGN',
-      campaign: paramsObject?.utm_campaign ?? '',
-      referrer: paramsObject?.utm_referrer ?? '',
-      medium: paramsObject?.utm_medium ?? '',
-      source: paramsObject?.utm_source ?? '',
+      campaign: referralObject?.utm_campaign ?? '',
+      referrer: referralObject?.utm_referrer ?? '',
+      medium: referralObject?.utm_medium ?? '',
+      source: referralObject?.utm_source ?? '',
       customer: {
         email: toggleAnonymous ? `${uuid.substring(0, 10)}@edustipend.org` : email,
         name: fullname,
